requirements.txt
Django==4.0.1
django-environ==0.8.1
djangorestframework==3.12.4 
psycopg2-binary==2.9.3
pytz==2022.1
django-filter==21.1
requests==2.27.1